Refactor structure into views and components
Refactored the previous components folder to separate
views (i.e. pages like the landing page and home page)
from components.

This will reduce file structure complexity while also
making it easier to reuse components.
