+++ client/src/components/DashboardIcon/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "DashboardIcon.js"
+}
