+++ client/src/components/Modals/CreateProject.js
@@ -1,52 +1,45 @@
 import React, { Component } from "react";
 import {
   Modal,
-  Form,
-  Input,
   Button,
   Icon,
+  Form,
+  Input,
   Divider,
   TextArea
 } from "semantic-ui-react";
-import store from "../../../store";
-import { hideEditProject, handleEditProject } from "../../../actions/index.js";
+import store from "../../store";
+import { hideCreateProject, handleAddProject } from "../../actions/index.js";
 
-class EditProject extends Component {
-  constructor(props) {
-    super(props);
-
+class CreateProject extends Component {
+  constructor() {
+    super();
     this.state = {
       show: false,
-      projectId: this.props.projectId,
-      projectName: this.props.projectName,
-      projectDescription: this.props.projectDescription,
-      project: {},
+      projectName: "",
+      projectDescription: "",
       usersMap: [],
-      users: [],
-      tasks: this.props.tasks,
-      category: ""
+      category: "My Projects",
+      userCategories: store.getState().handleProject.userCategories
     };
 
-    this.updateProject = this.updateProject.bind(this);
+    this.addProject = this.addProject.bind(this);
     this.handleProjectNameChange = this.handleProjectNameChange.bind(this);
     this.handleProjectDescriptionChange = this.handleProjectDescriptionChange.bind(
       this
     );
+    this.clearFields = this.clearFields.bind(this);
     this.handleClose = this.handleClose.bind(this);
   }
 
   componentDidMount() {
     this.unsubscribe = store.subscribe(() => {
       this.setState({
-        show: store.getState().editModalVisibility.visibility,
-        projectId: store.getState().editModalVisibility.projectId,
-        projectName: store.getState().editModalVisibility.projectName,
-        projectDescription: store.getState().editModalVisibility
-          .projectDescription,
-        project: store.getState().editModalVisibility.project,
-        usersMap: store.getState().editModalVisibility.usersMap,
-        tasks: store.getState().editModalVisibility.tasks,
-        category: store.getState().editModalVisibility.category
+        show: store.getState().createProject.visibility,
+        projectName: store.getState().createProject.projectName,
+        projectDescription: store.getState().createProject.projectDescription,
+        usersMap: store.getState().createProject.usersMap,
+        userCategories: store.getState().handleProject.userCategories
       });
     });
   }
@@ -55,40 +48,48 @@ class EditProject extends Component {
     this.unsubscribe();
   }
 
-  updateProject(event) {
+  addProject(event) {
     event.preventDefault();
     var users = [];
+    users.push(store.getState().userAuthentication.username);
     this.state.usersMap.map(user => {
       users.push(user.name);
     });
     fetch("/project-service/projects", {
-      method: "PUT",
+      method: "POST",
       headers: {
         Accept: "application/json",
         "Content-Type": "application/json",
         Authorization: "Bearer " + localStorage.getItem("JWT")
       },
       body: JSON.stringify({
-        projectId: this.state.projectId,
-        projectName: this.state.projectName,
-        projectDescription: this.state.projectDescription,
-        users: users,
-        tasks: this.state.tasks
-      })
-    }).then(response => {
-      if (response.status === 200) {
-        var payload = {
-          project: this.state.project,
+        project: {
           projectName: this.state.projectName,
           projectDescription: this.state.projectDescription,
           users: users,
-          category: this.state.category
-        };
-        store.dispatch(handleEditProject(payload));
+          tasks: []
+        },
+        projectCategory: this.state.category
+      })
+    })
+      .then(response => {
+        if (response.ok) {
+          return response.json();
+        } else {
+          throw new Error("Something went wrong ...");
+        }
+      })
+      .then(data => {
+        store.dispatch(handleAddProject({ data }));
         this.handleClose();
-      } else {
-        console.log("Error");
-      }
+      });
+  }
+
+  clearFields() {
+    this.setState({
+      projectName: "",
+      projectDescription: "",
+      usersMap: [{ name: "" }]
     });
   }
 
@@ -109,7 +110,8 @@ class EditProject extends Component {
     this.setState({ usersMap: newUsers });
   };
 
-  handleAddUser = () => {
+  handleAddUser = event => {
+    event.preventDefault();
     this.setState({
       usersMap: this.state.usersMap.concat([{ name: "" }])
     });
@@ -121,8 +123,31 @@ class EditProject extends Component {
     });
   };
 
+  handleCategoryChange = (event, result) => {
+    const { value } = result;
+    this.setState({
+      category: value
+    });
+  };
+
   handleClose() {
-    store.dispatch(hideEditProject());
+    store.dispatch(hideCreateProject());
+  }
+
+  getOptions() {
+    const { userCategories } = this.state;
+    const sortedCategories = ["My Projects", "Shared Projects"];
+    userCategories.sort().forEach(function(category) {
+      if (category !== "My Projects" && category !== "Shared Projects")
+        sortedCategories.push(category);
+    });
+
+    const options = sortedCategories.map(category => ({
+      key: category,
+      text: category,
+      value: category
+    }));
+    return options;
   }
 
   render() {
@@ -134,30 +159,43 @@ class EditProject extends Component {
         open={this.state.show}
         onClose={this.handleClose}
       >
-        <Modal.Header>Edit project ({this.state.projectId})</Modal.Header>
+        <Modal.Header>Create project</Modal.Header>
         <Modal.Content>
           <Form>
-            <Form.Field
-              label="Project Name"
-              control={Input}
-              placeholder="Enter Project Name"
-              onChange={this.handleProjectNameChange}
-              value={this.state.projectName}
-            />
+            <Form.Group>
+              <Form.Field
+                control={Input}
+                label="Project Name"
+                placeholder="Enter Project Name"
+                onChange={this.handleProjectNameChange}
+                width={10}
+              />
+              <Form.Select
+                fluid
+                label="Category"
+                options={this.getOptions()}
+                placeholder="My Projects"
+                defaultValue="My Projects"
+                width={6}
+                onChange={this.handleCategoryChange}
+              />
+            </Form.Group>
             <Divider />
             <Form.Field
-              label="Project Description"
               control={TextArea}
+              label="Project Description"
               placeholder="Enter Project Description"
               onChange={this.handleProjectDescriptionChange}
-              value={this.state.projectDescription}
             />
             <Divider />
+            <p>
+              <i>Note: You will automatically be added to the project.</i>
+            </p>
             {this.state.usersMap.map((user, idx) => (
-              <div className="edit-user">
+              <div className="modal-user" key={user.name}>
                 <Form.Group className="modal-row">
                   <Form.Input
-                    placeholder={"User #" + (idx + 1)}
+                    placeholder={"User #" + (idx + 2)}
                     value={user.name}
                     onChange={this.handleUserNameChange(idx)}
                     width={6}
@@ -181,8 +219,8 @@ class EditProject extends Component {
           <Button negative onClick={this.handleClose}>
             <Icon name="remove" /> Close
           </Button>
-          <Button positive type="submit" onClick={this.updateProject}>
-            <Icon name="checkmark" /> Save
+          <Button positive type="submit" onClick={this.addProject}>
+            <Icon name="checkmark" /> Submit
           </Button>
         </Modal.Actions>
       </Modal>
@@ -190,4 +228,4 @@ class EditProject extends Component {
   }
 }
 
-export default EditProject;
+export default CreateProject;
