+++ client/src/components/Modals/EditProject.js
@@ -1,45 +1,52 @@
 import React, { Component } from "react";
 import {
   Modal,
-  Button,
-  Icon,
   Form,
   Input,
+  Button,
+  Icon,
   Divider,
   TextArea
 } from "semantic-ui-react";
-import store from "../../../store";
-import { hideCreateProject, handleAddProject } from "../../../actions/index.js";
+import store from "../../store";
+import { hideEditProject, handleEditProject } from "../../actions/index.js";
 
-class CreateProject extends Component {
-  constructor() {
-    super();
+class EditProject extends Component {
+  constructor(props) {
+    super(props);
+
     this.state = {
       show: false,
-      projectName: "",
-      projectDescription: "",
+      projectId: this.props.projectId,
+      projectName: this.props.projectName,
+      projectDescription: this.props.projectDescription,
+      project: {},
       usersMap: [],
-      category: "My Projects",
-      userCategories: store.getState().handleProject.userCategories
+      users: [],
+      tasks: this.props.tasks,
+      category: ""
     };
 
-    this.addProject = this.addProject.bind(this);
+    this.updateProject = this.updateProject.bind(this);
     this.handleProjectNameChange = this.handleProjectNameChange.bind(this);
     this.handleProjectDescriptionChange = this.handleProjectDescriptionChange.bind(
       this
     );
-    this.clearFields = this.clearFields.bind(this);
     this.handleClose = this.handleClose.bind(this);
   }
 
   componentDidMount() {
     this.unsubscribe = store.subscribe(() => {
       this.setState({
-        show: store.getState().createProject.visibility,
-        projectName: store.getState().createProject.projectName,
-        projectDescription: store.getState().createProject.projectDescription,
-        usersMap: store.getState().createProject.usersMap,
-        userCategories: store.getState().handleProject.userCategories
+        show: store.getState().editModalVisibility.visibility,
+        projectId: store.getState().editModalVisibility.projectId,
+        projectName: store.getState().editModalVisibility.projectName,
+        projectDescription: store.getState().editModalVisibility
+          .projectDescription,
+        project: store.getState().editModalVisibility.project,
+        usersMap: store.getState().editModalVisibility.usersMap,
+        tasks: store.getState().editModalVisibility.tasks,
+        category: store.getState().editModalVisibility.category
       });
     });
   }
@@ -48,48 +55,40 @@ class CreateProject extends Component {
     this.unsubscribe();
   }
 
-  addProject(event) {
+  updateProject(event) {
     event.preventDefault();
     var users = [];
-    users.push(store.getState().userAuthentication.username);
     this.state.usersMap.map(user => {
       users.push(user.name);
     });
     fetch("/project-service/projects", {
-      method: "POST",
+      method: "PUT",
       headers: {
         Accept: "application/json",
         "Content-Type": "application/json",
         Authorization: "Bearer " + localStorage.getItem("JWT")
       },
       body: JSON.stringify({
-        project: {
+        projectId: this.state.projectId,
+        projectName: this.state.projectName,
+        projectDescription: this.state.projectDescription,
+        users: users,
+        tasks: this.state.tasks
+      })
+    }).then(response => {
+      if (response.status === 200) {
+        var payload = {
+          project: this.state.project,
           projectName: this.state.projectName,
           projectDescription: this.state.projectDescription,
           users: users,
-          tasks: []
-        },
-        projectCategory: this.state.category
-      })
-    })
-      .then(response => {
-        if (response.ok) {
-          return response.json();
-        } else {
-          throw new Error("Something went wrong ...");
-        }
-      })
-      .then(data => {
-        store.dispatch(handleAddProject({ data }));
+          category: this.state.category
+        };
+        store.dispatch(handleEditProject(payload));
         this.handleClose();
-      });
-  }
-
-  clearFields() {
-    this.setState({
-      projectName: "",
-      projectDescription: "",
-      usersMap: [{ name: "" }]
+      } else {
+        console.log("Error");
+      }
     });
   }
 
@@ -110,8 +109,7 @@ class CreateProject extends Component {
     this.setState({ usersMap: newUsers });
   };
 
-  handleAddUser = event => {
-    event.preventDefault();
+  handleAddUser = () => {
     this.setState({
       usersMap: this.state.usersMap.concat([{ name: "" }])
     });
@@ -123,31 +121,8 @@ class CreateProject extends Component {
     });
   };
 
-  handleCategoryChange = (event, result) => {
-    const { value } = result;
-    this.setState({
-      category: value
-    });
-  };
-
   handleClose() {
-    store.dispatch(hideCreateProject());
-  }
-
-  getOptions() {
-    const { userCategories } = this.state;
-    const sortedCategories = ["My Projects", "Shared Projects"];
-    userCategories.sort().forEach(function(category) {
-      if (category !== "My Projects" && category !== "Shared Projects")
-        sortedCategories.push(category);
-    });
-
-    const options = sortedCategories.map(category => ({
-      key: category,
-      text: category,
-      value: category
-    }));
-    return options;
+    store.dispatch(hideEditProject());
   }
 
   render() {
@@ -159,43 +134,30 @@ class CreateProject extends Component {
         open={this.state.show}
         onClose={this.handleClose}
       >
-        <Modal.Header>Create project</Modal.Header>
+        <Modal.Header>Edit project ({this.state.projectId})</Modal.Header>
         <Modal.Content>
           <Form>
-            <Form.Group>
-              <Form.Field
-                control={Input}
-                label="Project Name"
-                placeholder="Enter Project Name"
-                onChange={this.handleProjectNameChange}
-                width={10}
-              />
-              <Form.Select
-                fluid
-                label="Category"
-                options={this.getOptions()}
-                placeholder="My Projects"
-                defaultValue="My Projects"
-                width={6}
-                onChange={this.handleCategoryChange}
-              />
-            </Form.Group>
-            <Divider />
             <Form.Field
-              control={TextArea}
-              label="Project Description"
-              placeholder="Enter Project Description"
-              onChange={this.handleProjectDescriptionChange}
+              label="Project Name"
+              control={Input}
+              placeholder="Enter Project Name"
+              onChange={this.handleProjectNameChange}
+              value={this.state.projectName}
             />
             <Divider />
-            <p>
-              <i>Note: You will automatically be added to the project.</i>
-            </p>
+            <Form.Field
+              label="Project Description"
+              control={TextArea}
+              placeholder="Enter Project Description"
+              onChange={this.handleProjectDescriptionChange}
+              value={this.state.projectDescription}
+            />
+            <Divider />
             {this.state.usersMap.map((user, idx) => (
-              <div className="modal-user" key={user.name}>
+              <div className="edit-user">
                 <Form.Group className="modal-row">
                   <Form.Input
-                    placeholder={"User #" + (idx + 2)}
+                    placeholder={"User #" + (idx + 1)}
                     value={user.name}
                     onChange={this.handleUserNameChange(idx)}
                     width={6}
@@ -219,8 +181,8 @@ class CreateProject extends Component {
           <Button negative onClick={this.handleClose}>
             <Icon name="remove" /> Close
           </Button>
-          <Button positive type="submit" onClick={this.addProject}>
-            <Icon name="checkmark" /> Submit
+          <Button positive type="submit" onClick={this.updateProject}>
+            <Icon name="checkmark" /> Save
           </Button>
         </Modal.Actions>
       </Modal>
@@ -228,4 +190,4 @@ class CreateProject extends Component {
   }
 }
 
-export default CreateProject;
+export default EditProject;
