+++ client/src/components/NewsItem/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "NewsItem.js"
+}
