+++ client/src/components/Project/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "Project.js"
+}
