+++ client/src/components/ProjectDropdown/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "ProjectDropdown.js"
+}
