+++ client/src/components/Sidebar/SidebarProject.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import store from "../../store";
 import { goToProject } from "../../actions";
 import { Menu } from "semantic-ui-react";
-import ProjectDropdown from "../ProjectsPage/Projects/Dropdown/ProjectDropdown";
+import ProjectDropdown from "../ProjectDropdown";
 
 export default class SidebarProject extends Component {
   getIndex = (category, projectId) => {
