+++ client/src/views/RegisterPage/package.json
